Pull the latest change to my fork

+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/json/WxMpUserGsonAdapter.java
@@ -8,16 +8,11 @@
  */
 package me.chanjar.weixin.mp.util.json;
 
-import java.lang.reflect.Type;
-
-import com.google.gson.JsonDeserializationContext;
-import com.google.gson.JsonDeserializer;
-import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParseException;
-
+import com.google.gson.*;
 import me.chanjar.weixin.common.util.json.GsonHelper;
 import me.chanjar.weixin.mp.bean.result.WxMpUser;
+
+import java.lang.reflect.Type;
 
 public class WxMpUserGsonAdapter implements JsonDeserializer<WxMpUser> {
 
@@ -41,9 +36,9 @@ public class WxMpUserGsonAdapter implements JsonDeserializer<WxMpUser> {
     Integer sexId = GsonHelper.getInteger(o, "sex");
     wxMpUser.setRemark(GsonHelper.getString(o, "remark"));
     wxMpUser.setGroupId(GsonHelper.getInteger(o, "groupid"));
-    wxMpUser.setTagIds(GsonHelper.getIntArray(o, "tagid_list"));
+    wxMpUser.setTagIds(GsonHelper.getLongArray(o, "tagid_list"));
     wxMpUser.setSexId(sexId);
-    if(new Integer(1).equals(sexId)) {
+    if (new Integer(1).equals(sexId)) {
       wxMpUser.setSex("男");
     } else if (new Integer(2).equals(sexId)) {
       wxMpUser.setSex("女");
